Refactor and simplify ultimate redirect hunter to final_hunter.py
