+++ report.html
@@ -1,0 +1,68 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>🔥 ULTIMATE HUNTER REPORT 🔥</title>
+    <style>
+        body {
+            font-family: 'Courier New', monospace;
+            background: #0a0a0a;
+            color: #00ff41;
+            margin: 20px;
+        }
+        .container {
+            max-width: 1200px;
+            margin: 0 auto;
+            background: #000;
+            border: 2px solid #00ff41;
+            padding: 30px;
+            border-radius: 10px;
+        }
+        .header {
+            text-align: center;
+            border-bottom: 2px solid #00ff41;
+            padding-bottom: 20px;
+            margin-bottom: 30px;
+        }
+        .vulnerability {
+            background: #1a0000;
+            border: 1px solid #ff4444;
+            padding: 20px;
+            margin: 20px 0;
+            border-radius: 8px;
+        }
+        .parameter {
+            background: #001a00;
+            border: 1px solid #00ff41;
+            padding: 15px;
+            margin: 15px 0;
+            border-radius: 6px;
+        }
+        .code {
+            background: #000;
+            color: #00ff41;
+            padding: 10px;
+            border: 1px solid #00ff41;
+            font-family: monospace;
+        }
+        .success { color: #00ff41; font-weight: bold; }
+        .error { color: #ff4444; font-weight: bold; }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <div class="header">
+            <h1>🔥 ULTIMATE REDIRECT HUNTER REPORT 🔥</h1>
+            <p>Target: https://httpbin.org/get?url=test&redirect=example</p>
+            <p>Scan Date: 2025-09-03 22:39:52</p>
+        </div>
+        
+        <h2>📊 SCAN SUMMARY</h2>
+        <p>URLs Scanned: 1</p>
+        <p>Parameters Found: 2</p>
+        <p>Redirect Parameters: 2</p>
+        <p>Vulnerabilities: <span class="success">0</span></p>
+
+    </div>
+</body>
+</html>
