+++ results.json
@@ -1,0 +1,31 @@
+{
+  "scan_info": {
+    "target_url": "https://httpbin.org/get?url=test&redirect=example",
+    "scan_date": "2025-09-03T22:39:52.370203",
+    "total_parameters": 2,
+    "vulnerabilities_found": 0
+  },
+  "parameters": [
+    {
+      "name": "url",
+      "value": "test",
+      "source": "url",
+      "context": "query",
+      "url": "https://httpbin.org/get?url=test&redirect=example",
+      "method": "GET",
+      "is_redirect_related": true,
+      "confidence": 0.8999999999999999
+    },
+    {
+      "name": "redirect",
+      "value": "example",
+      "source": "url",
+      "context": "query",
+      "url": "https://httpbin.org/get?url=test&redirect=example",
+      "method": "GET",
+      "is_redirect_related": true,
+      "confidence": 0.8999999999999999
+    }
+  ],
+  "vulnerabilities": []
+}
